Refactor: Split oversized test files into smaller parts
Splits the following test files that exceeded 300 SLOCs into
more manageable, logically grouped files:

- tests/core/test_quota_service.py
- tests/accounting/test_account_model_limits.py
- tests/test_migrations.py

This is a structural refactoring only, intended to improve
maintainability. Functional changes were not made.

Note: Sandbox environment issues prevented test execution and
coverage checks after refactoring.
