+++ tests/core/test_quota_service_check_quota_enhanced.py
@@ -1,13 +1,15 @@
 from datetime import datetime, timedelta, timezone
-from unittest.mock import MagicMock, call, patch
+from unittest.mock import MagicMock, patch
 import pytest
-from typing import Optional, Dict, Tuple
+from typing import Optional
+
+from freezegun import freeze_time
 
 from llm_accounting.models.limits import (LimitScope, LimitType, TimeInterval,
                                           UsageLimitDTO)
 from llm_accounting.services.quota_service import QuotaService
 from llm_accounting.backends.base import TransactionalBackend
-from llm_accounting import LLMAccounting # Added import
+
 
 @pytest.fixture
 def mock_backend() -> MagicMock:
@@ -15,310 +17,6 @@ def mock_backend() -> MagicMock:
     backend = MagicMock(spec=TransactionalBackend)
     backend.get_usage_limits.return_value = []
     return backend
-
-
-def test_check_quota_no_limits(mock_backend: MagicMock):
-    """Test check_quota when no limits are configured (cache is empty)."""
-    quota_service = QuotaService(mock_backend)
-
-    is_allowed, reason = quota_service.check_quota(
-        model="gpt-4", username="test_user", caller_name="test_caller",
-        input_tokens=100, cost=0.01
-    )
-    
-    assert is_allowed is True
-    assert reason is None
-    mock_backend.get_usage_limits.assert_called_once()
-
-
-def test_check_quota_allowed_single_limit(mock_backend: MagicMock):
-    """Test check_quota when usage is within a single configured limit."""
-    now = datetime.now(timezone.utc)
-    user_cost_limit = UsageLimitDTO(
-        id=1, scope=LimitScope.USER.value, limit_type=LimitType.COST.value,
-        max_value=10.0, interval_unit=TimeInterval.MONTH.value, interval_value=1,
-        username="test_user", created_at=now, updated_at=now
-    )
-    mock_backend.get_usage_limits.return_value = [user_cost_limit]
-    quota_service = QuotaService(mock_backend)
-
-    mock_backend.get_accounting_entries_for_quota.return_value = 5.0 
-    
-    is_allowed, reason = quota_service.check_quota(
-        model="gpt-4", username="test_user", caller_name="test_caller",
-        input_tokens=100, cost=0.01
-    )
-    
-    assert is_allowed is True
-    assert reason is None
-
-    mock_backend.get_usage_limits.assert_called_once()
-    mock_backend.get_accounting_entries_for_quota.assert_called_once()
-    kwargs = mock_backend.get_accounting_entries_for_quota.call_args.kwargs
-    assert kwargs['limit_type'] == LimitType.COST
-    assert kwargs['username'] == "test_user"
-
-
-def test_check_quota_denied_single_limit(mock_backend: MagicMock):
-    """Test check_quota when usage exceeds a single configured limit."""
-    now = datetime.now(timezone.utc)
-    user_cost_limit = UsageLimitDTO(
-        id=1, scope=LimitScope.USER.value, limit_type=LimitType.COST.value,
-        max_value=10.0, interval_unit=TimeInterval.MONTH.value, interval_value=1,
-        username="test_user", created_at=now, updated_at=now
-    )
-    mock_backend.get_usage_limits.return_value = [user_cost_limit]
-    # Instantiate QuotaService AFTER setting the mock return value.
-    # The first call to check_quota will load the cache if it's None.
-    quota_service = QuotaService(mock_backend)
-
-    mock_backend.get_accounting_entries_for_quota.return_value = 9.99
-    
-    is_allowed, reason = quota_service.check_quota(
-        model="gpt-4", username="test_user", caller_name="test_caller",
-        input_tokens=0, cost=0.02
-    )
-    
-    assert is_allowed is False
-    assert reason is not None
-    assert "USER (user: test_user) limit: 10.00 cost per 1 month" in reason
-    assert "exceeded. Current usage: 9.99, request: 0.02." in reason
-
-    mock_backend.get_usage_limits.assert_called_once()
-    mock_backend.get_accounting_entries_for_quota.assert_called_once()
-
-
-def test_check_quota_multiple_limits_one_exceeded(mock_backend: MagicMock):
-    """Test check_quota with multiple limits, where one is exceeded."""
-    now = datetime.now(timezone.utc)
-    cost_limit_user = UsageLimitDTO(
-        id=1, scope=LimitScope.USER.value, limit_type=LimitType.COST.value,
-        max_value=10.0, interval_unit=TimeInterval.MONTH.value, interval_value=1,
-        username="test_user", created_at=now, updated_at=now
-    )
-    request_limit_user = UsageLimitDTO(
-        id=2, scope=LimitScope.USER.value, limit_type=LimitType.REQUESTS.value,
-        max_value=100.0, interval_unit=TimeInterval.DAY.value, interval_value=1,
-        username="test_user", created_at=now, updated_at=now
-    )
-    mock_backend.get_usage_limits.return_value = [cost_limit_user, request_limit_user]
-    quota_service = QuotaService(mock_backend)
-
-    def get_accounting_side_effect(start_time, end_time, limit_type, interval_unit, model, username, caller_name, project_name, filter_project_null):
-        if limit_type == LimitType.COST and username == "test_user":
-            return 5.0
-        elif limit_type == LimitType.REQUESTS and username == "test_user":
-            return 100.0
-        return 0.0
-    
-    mock_backend.get_accounting_entries_for_quota.side_effect = get_accounting_side_effect
-    
-    is_allowed, reason = quota_service.check_quota(
-        model="gpt-4", username="test_user", caller_name="test_caller",
-        input_tokens=10, cost=0.01
-    )
-    
-    assert is_allowed is False
-    assert reason is not None
-    assert "USER (user: test_user) limit: 100.00 requests per 1 day" in reason
-    assert "exceeded. Current usage: 100.00, request: 1.00." in reason
-
-    mock_backend.get_usage_limits.assert_called_once()
-    assert mock_backend.get_accounting_entries_for_quota.call_count == 2
-
-
-def test_check_quota_different_scopes_in_cache(mock_backend: MagicMock):
-    """Test that QuotaService correctly filters from cache for different scopes."""
-    now = datetime.now(timezone.utc)
-    global_req_limit = UsageLimitDTO(id=1, scope=LimitScope.GLOBAL.value, limit_type=LimitType.REQUESTS.value, max_value=5, interval_unit=TimeInterval.MINUTE.value, interval_value=1)
-    user_cost_limit = UsageLimitDTO(id=2, scope=LimitScope.USER.value, username="test_user", limit_type=LimitType.COST.value, max_value=10, interval_unit=TimeInterval.DAY.value, interval_value=1)
-    model_token_limit = UsageLimitDTO(id=3, scope=LimitScope.MODEL.value, model="gpt-4", limit_type=LimitType.INPUT_TOKENS.value, max_value=1000, interval_unit=TimeInterval.HOUR.value, interval_value=1)
-    
-    mock_backend.get_usage_limits.return_value = [global_req_limit, user_cost_limit, model_token_limit]
-    quota_service = QuotaService(mock_backend)
-    
-    mock_backend.get_accounting_entries_for_quota.return_value = 5.0
-
-    is_allowed, reason = quota_service.check_quota(
-        model="gpt-4", username="test_user", caller_name="super_caller",
-        input_tokens=1, cost=0.001
-    )
-    
-    assert not is_allowed
-    assert "GLOBAL limit: 5.00 requests per 1 minute" in reason
-
-    mock_backend.get_usage_limits.assert_called_once()
-    assert mock_backend.get_accounting_entries_for_quota.call_count == 3
-    call_kwargs_list = [call.kwargs for call in mock_backend.get_accounting_entries_for_quota.call_args_list]
-    assert any(kw['limit_type'] == LimitType.REQUESTS and kw['model'] is None and kw['username'] is None for kw in call_kwargs_list)
-
-
-def test_check_quota_token_limits(mock_backend: MagicMock):
-    """Test check_quota for input token limits from cache."""
-    now = datetime.now(timezone.utc)
-    model_token_limit = UsageLimitDTO(
-        id=1, scope=LimitScope.MODEL.value, limit_type=LimitType.INPUT_TOKENS.value,
-        max_value=1000.0, interval_unit=TimeInterval.HOUR.value, interval_value=1,
-        model="text-davinci-003", created_at=now, updated_at=now
-    )
-    mock_backend.get_usage_limits.return_value = [model_token_limit]
-    quota_service = QuotaService(mock_backend)
-
-    mock_backend.get_accounting_entries_for_quota.return_value = 950.0
-
-    is_allowed, reason = quota_service.check_quota(
-        model="text-davinci-003", username="any_user", caller_name="any_caller",
-        input_tokens=50, cost=0.0
-    )
-    assert is_allowed is True
-    assert reason is None
-    mock_backend.get_accounting_entries_for_quota.assert_called_with(
-        start_time=mock_backend.get_accounting_entries_for_quota.call_args.kwargs['start_time'],
-        end_time=mock_backend.get_accounting_entries_for_quota.call_args.kwargs['end_time'], # Add end_time
-        limit_type=LimitType.INPUT_TOKENS,
-        interval_unit=mock_backend.get_accounting_entries_for_quota.call_args.kwargs['interval_unit'], # Add interval_unit
-        model="text-davinci-003",
-        username=None, caller_name=None, project_name=None, filter_project_null=None
-    )
-
-    mock_backend.get_accounting_entries_for_quota.reset_mock()
-    mock_backend.get_accounting_entries_for_quota.return_value = 950.0
-
-    is_allowed, reason = quota_service.check_quota(
-        model="text-davinci-003", username="any_user", caller_name="any_caller",
-        input_tokens=51, cost=0.0
-    )
-    assert is_allowed is False
-    assert reason is not None
-    assert "MODEL (model: text-davinci-003) limit: 1000.00 input_tokens per 1 hour" in reason
-    assert "exceeded. Current usage: 950.00, request: 51.00." in reason
-
-    mock_backend.get_usage_limits.assert_called_once()
-    assert mock_backend.get_accounting_entries_for_quota.call_count == 1
-
-
-def test_check_quota_total_token_limits(mock_backend: MagicMock):
-    """Test check_quota for total token limits."""
-    now = datetime.now(timezone.utc)
-    total_token_limit = UsageLimitDTO(
-        id=1,
-        scope=LimitScope.USER.value,
-        limit_type=LimitType.TOTAL_TOKENS.value,
-        max_value=500.0,
-        interval_unit=TimeInterval.DAY.value,
-        interval_value=1,
-        username="user_total",
-        created_at=now,
-        updated_at=now,
-    )
-    mock_backend.get_usage_limits.return_value = [total_token_limit]
-    quota_service = QuotaService(mock_backend)
-
-    mock_backend.get_accounting_entries_for_quota.return_value = 480.0
-
-    is_allowed, reason = quota_service.check_quota(
-        model="model-a",
-        username="user_total",
-        caller_name="caller",
-        input_tokens=10,
-        cost=0.0,
-        completion_tokens=5,
-    )
-    assert is_allowed is True
-    assert reason is None
-
-    mock_backend.get_accounting_entries_for_quota.reset_mock()
-    mock_backend.get_accounting_entries_for_quota.return_value = 490.0
-
-    is_allowed, reason = quota_service.check_quota(
-        model="model-a",
-        username="user_total",
-        caller_name="caller",
-        input_tokens=15,
-        cost=0.0,
-        completion_tokens=10,
-    )
-    assert not is_allowed
-    assert reason is not None
-    assert "USER (user: user_total) limit: 500.00 total_tokens per 1 day" in reason
-    assert "exceeded. Current usage: 490.00, request: 25.00." in reason
-
-    mock_backend.get_usage_limits.assert_called()
-    assert mock_backend.get_accounting_entries_for_quota.call_count == 1
-
-
-def test_get_period_start_monthly(mock_backend: MagicMock):
-    quota_service = QuotaService(mock_backend)
-    current_time = datetime(2024, 3, 15, 10, 30, 0, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.MONTH, 1)
-    assert period_start == datetime(2024, 3, 1, 0, 0, 0, tzinfo=timezone.utc)
-
-    current_time = datetime(2024, 4, 1, 0, 0, 0, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.MONTH, 1)
-    assert period_start == datetime(2024, 4, 1, 0, 0, 0, tzinfo=timezone.utc)
-
-def test_get_period_start_daily(mock_backend: MagicMock):
-    quota_service = QuotaService(mock_backend)
-    current_time = datetime(2024, 3, 15, 10, 30, 0, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.DAY, 1)
-    assert period_start == datetime(2024, 3, 15, 0, 0, 0, tzinfo=timezone.utc)
-
-def test_get_period_start_hourly(mock_backend: MagicMock):
-    quota_service = QuotaService(mock_backend)
-    current_time = datetime(2024, 3, 15, 10, 30, 0, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.HOUR, 1)
-    assert period_start == datetime(2024, 3, 15, 10, 0, 0, tzinfo=timezone.utc)
-
-def test_get_period_start_minute(mock_backend: MagicMock):
-    quota_service = QuotaService(mock_backend)
-    current_time = datetime(2024, 3, 15, 10, 37, 45, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.MINUTE, 1)
-    assert period_start == datetime(2024, 3, 15, 10, 37, 0, tzinfo=timezone.utc)
-
-    current_time = datetime(2024, 3, 15, 10, 37, 45, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.MINUTE, 5)
-    assert period_start == datetime(2024, 3, 15, 10, 35, 0, tzinfo=timezone.utc)
-
-def test_get_period_start_second(mock_backend: MagicMock):
-    quota_service = QuotaService(mock_backend)
-    current_time = datetime(2024, 3, 15, 10, 37, 45, 123456, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.SECOND, 1)
-    assert period_start == datetime(2024, 3, 15, 10, 37, 45, 0, tzinfo=timezone.utc)
-
-    current_time = datetime(2024, 3, 15, 10, 37, 45, 123456, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.SECOND, 10)
-    assert period_start == datetime(2024, 3, 15, 10, 37, 40, 0, tzinfo=timezone.utc)
-
-def test_get_period_start_weekly(mock_backend: MagicMock):
-    quota_service = QuotaService(mock_backend)
-    current_time = datetime(2024, 3, 13, 10, 30, 0, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.WEEK, 1)
-    assert period_start == datetime(2024, 3, 11, 0, 0, 0, tzinfo=timezone.utc)
-
-    current_time = datetime(2024, 3, 11, 10, 30, 0, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.WEEK, 1)
-    assert period_start == datetime(2024, 3, 11, 0, 0, 0, tzinfo=timezone.utc)
-
-    current_time = datetime(2024, 3, 11, 10, 30, 0, tzinfo=timezone.utc)
-    period_start = quota_service.limit_evaluator._get_period_start(current_time, TimeInterval.WEEK, 2)
-    assert period_start == datetime(2024, 3, 4, 0, 0, 0, tzinfo=timezone.utc)
-
-def test_get_period_start_unsupported_interval(mock_backend: MagicMock):
-    quota_service = QuotaService(mock_backend)
-    current_time = datetime.now(timezone.utc)
-    # Test that attempting to create a TimeInterval enum from an invalid string raises ValueError
-    with pytest.raises(ValueError, match="'unsupported_unit' is not a valid TimeInterval"):
-        TimeInterval("unsupported_unit")
-    # Further test: if _get_period_start was somehow called with a non-enum (e.g. direct test),
-    # it should also fail. This part is commented out as the primary test above is sufficient
-    # for typical usage where enum conversion happens before calling _get_period_start.
-    # The original test was trying to test _get_period_start's robustness to incorrect types,
-    # which resulted in AttributeError. The more relevant test is the Enum conversion itself.
-    # If direct calls to _get_period_start with strings were a supported scenario,
-    # then _get_period_start would need explicit type checking.
-    #
-    # with pytest.raises(AttributeError): # Or appropriate error if type check added
-    #     quota_service.limit_evaluator._get_period_start(current_time, "unsupported_unit", 1)
 
 
 # --- Tests for check_quota_enhanced ---
@@ -362,8 +60,6 @@ def test_check_quota_enhanced_allowed_single_limit(mock_backend: MagicMock):
     mock_backend.get_usage_limits.assert_called_once()
     mock_backend.get_accounting_entries_for_quota.assert_called_once()
 
-
-from freezegun import freeze_time
 
 # Helper to calculate expected retry_after based on reset_timestamp
 def calculate_expected_retry_after(reset_timestamp: Optional[datetime], current_time: datetime) -> Optional[int]:
@@ -697,98 +393,3 @@ def test_check_quota_enhanced_cache_expires(mock_backend: MagicMock):
             assert retry_after_expired is None # retry_after should be None for allowed
             mock_evaluate_enhanced.assert_called_once()
             assert cache_key not in quota_service._denial_cache # Assert cache entry is gone
-
-
-def test_cache_rebuild_after_inserting_limit(memory_sqlite_backend):
-    accounting = LLMAccounting(backend=memory_sqlite_backend)
-    # Define request parameters
-    model = "test-model-insert"
-    username = "test-user-insert"
-    caller_name = "test-caller-insert"
-    project_name = "test-project-insert"
-    input_tokens = 100
-    cost = 1.0
-
-    # 1. Initial check (should be allowed)
-    allowed, reason = accounting.check_quota(
-        model=model,
-        username=username,
-        caller_name=caller_name,
-        project_name=project_name,
-        input_tokens=input_tokens,
-        cost=cost
-    )
-    assert allowed is True, f"Initial check should be allowed, but was denied: {reason}"
-
-    # 2. Insert a restrictive limit
-    accounting.set_usage_limit(
-        scope=LimitScope.USER,
-        limit_type=LimitType.COST,
-        max_value=0.5, # Restrictive limit
-        interval_unit=TimeInterval.DAY,
-        interval_value=1,
-        username=username
-    )
-
-    # 3. Check again (should be denied due to new limit)
-    allowed, reason = accounting.check_quota(
-        model=model,
-        username=username,
-        caller_name=caller_name,
-        project_name=project_name,
-        input_tokens=input_tokens,
-        cost=cost
-    )
-    assert allowed is False, "Check after inserting limit should be denied"
-    assert reason is not None, "Reason for denial should not be None"
-    # Example of a more specific reason check, adapt if necessary
-    assert f"USER (user: {username}) limit: 0.50 cost per 1 day exceeded." in reason
-
-    # 4. Clean up: Find and delete the limit
-    limits = accounting.get_usage_limits(username=username, scope=LimitScope.USER)
-    limit_deleted = False
-    for limit in limits:
-        if limit.limit_type == LimitType.COST.value and limit.max_value == 0.5 and limit.username == username: # Identify the specific limit
-            accounting.delete_usage_limit(limit.id)
-            limit_deleted = True
-            break
-    assert limit_deleted, "Could not find the test limit to delete"
-
-
-def test_cache_rebuild_after_deleting_limit(memory_sqlite_backend):
-    accounting = LLMAccounting(backend=memory_sqlite_backend)
-    model = "test-model-deletion"
-    username = "test-user-deletion"
-    caller_name = "test-caller-deletion"
-    project_name = "test-project-deletion"
-    input_tokens = 10
-    cost = 0.1
-
-    # 1. Insert a restrictive limit first
-    accounting.set_usage_limit(
-        scope=LimitScope.USER,
-        limit_type=LimitType.INPUT_TOKENS,
-        max_value=5, # Restrictive
-        interval_unit=TimeInterval.HOUR,
-        interval_value=1,
-        username=username
-    )
-
-    # Verify it's active
-    allowed, reason = accounting.check_quota(model, username, caller_name, input_tokens, cost, project_name=project_name)
-    assert allowed is False, f"Request should be denied by the new limit: {reason}"
-
-    # 2. Find and delete the limit
-    limits = accounting.get_usage_limits(username=username, scope=LimitScope.USER)
-    limit_id_to_delete = None
-    for limit_obj in limits:
-        if limit_obj.limit_type == LimitType.INPUT_TOKENS.value and limit_obj.max_value == 5 and limit_obj.username == username:
-            limit_id_to_delete = limit_obj.id
-            break
-    assert limit_id_to_delete is not None, "Test limit was not found for deletion"
-
-    accounting.delete_usage_limit(limit_id_to_delete)
-
-    # 3. Check again (should be allowed as the limit is gone)
-    allowed, reason = accounting.check_quota(model, username, caller_name, input_tokens, cost, project_name=project_name)
-    assert allowed is True, f"Request should be allowed after deleting limit, but was denied: {reason}"
